drop dependency on const_value from model
